decoupled backend api reference from cli, destructured cli file to seperate concerns for comparing algorithms and input types
