+++ src/python/cli/cli_config.py
@@ -1,0 +1,1 @@
+API_URL = "http://127.0.0.1:8000"
