--- src/python/gui/config.py
@@ -1,3 +1,0 @@
-API_URL = "http://127.0.0.1:8000"
-
-config = {}
