+++ src/python/gui/gui_config.py
@@ -1,0 +1,4 @@
+API_URL = "http://127.0.0.1:8000"
+
+# to be populated by http get request at runtime
+config = {}
