+++ up.sh
@@ -1,0 +1,1 @@
+fastapi dev src/python/backend/main.py
